backend/requirements.txt
asgiref==3.2.10
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
cryptography==3.0
Django==3.0.8
django-cors-headers==3.4.0
django-furl==0.1.0
django-rest-framework==0.1.0
django-rest-knox==4.1.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
furl==2.1.0
idna==2.10
orderedmultidict==1.0.1
Pillow==7.2.0
psycopg2==2.8.5
pycparser==2.20
PyJWT==1.7.1
python-dotenv==0.15.0
pytz==2020.1
requests==2.24.0
six==1.15.0
spotipy==2.13.0
sqlparse==0.3.1
urllib3==1.25.10
